exceptions moved to a new package
